+++ pkg/api/query/v1.go
@@ -44,9 +44,9 @@ import (
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/util/annotations"
 	"github.com/prometheus/prometheus/util/stats"
-	v1 "github.com/prometheus/prometheus/web/api/v1"
 	promqlapi "github.com/thanos-io/promql-engine/api"
 	"github.com/thanos-io/promql-engine/engine"
+	"github.com/thanos-io/promql-engine/logicalplan"
 
 	"github.com/thanos-io/thanos/pkg/api"
 	"github.com/thanos-io/thanos/pkg/exemplars"
@@ -90,19 +90,25 @@ const (
 	PromqlEngineThanos     PromqlEngineType = "thanos"
 )
 
+type ThanosEngine interface {
+	promql.QueryEngine
+	NewInstantQueryFromPlan(ctx context.Context, q storage.Queryable, opts promql.QueryOpts, plan logicalplan.Node, ts time.Time) (promql.Query, error)
+	NewRangeQueryFromPlan(ctx context.Context, q storage.Queryable, opts promql.QueryOpts, root logicalplan.Node, start, end time.Time, step time.Duration) (promql.Query, error)
+}
+
 type QueryEngineFactory struct {
 	engineOpts            promql.EngineOpts
 	remoteEngineEndpoints promqlapi.RemoteEndpoints
 
 	createPrometheusEngine sync.Once
-	prometheusEngine       v1.QueryEngine
+	prometheusEngine       promql.QueryEngine
 
 	createThanosEngine sync.Once
-	thanosEngine       v1.QueryEngine
+	thanosEngine       ThanosEngine
 	enableXFunctions   bool
 }
 
-func (f *QueryEngineFactory) GetPrometheusEngine() v1.QueryEngine {
+func (f *QueryEngineFactory) GetPrometheusEngine() promql.QueryEngine {
 	f.createPrometheusEngine.Do(func() {
 		if f.prometheusEngine != nil {
 			return
@@ -113,7 +119,7 @@ func (f *QueryEngineFactory) GetPrometheusEngine() v1.QueryEngine {
 	return f.prometheusEngine
 }
 
-func (f *QueryEngineFactory) GetThanosEngine() v1.QueryEngine {
+func (f *QueryEngineFactory) GetThanosEngine() ThanosEngine {
 	f.createThanosEngine.Do(func() {
 		if f.thanosEngine != nil {
 			return
@@ -305,6 +311,8 @@ type queryTelemetry struct {
 	// TODO(saswatamcode): Add aggregate fields to enrich data.
 	OperatorName string           `json:"name,omitempty"`
 	Execution    string           `json:"executionTime,omitempty"`
+	PeakSamples  int64            `json:"peakSamples,omitempty"`
+	TotalSamples int64            `json:"totalSamples,omitempty"`
 	Children     []queryTelemetry `json:"children,omitempty"`
 }
 
@@ -321,8 +329,8 @@ func (qapi *QueryAPI) parseEnableDedupParam(r *http.Request) (enableDeduplicatio
 	return enableDeduplication, nil
 }
 
-func (qapi *QueryAPI) parseEngineParam(r *http.Request) (queryEngine v1.QueryEngine, e PromqlEngineType, _ *api.ApiError) {
-	var engine v1.QueryEngine
+func (qapi *QueryAPI) parseEngineParam(r *http.Request) (queryEngine promql.QueryEngine, e PromqlEngineType, _ *api.ApiError) {
+	var engine promql.QueryEngine
 
 	param := PromqlEngineType(r.FormValue("engine"))
 	if param == "" {
@@ -471,8 +479,10 @@ func (qapi *QueryAPI) parseQueryAnalyzeParam(r *http.Request, query promql.Query
 
 func processAnalysis(a *engine.AnalyzeOutputNode) queryTelemetry {
 	var analysis queryTelemetry
-	analysis.OperatorName = a.OperatorTelemetry.Name()
+	analysis.OperatorName = a.OperatorTelemetry.String()
 	analysis.Execution = a.OperatorTelemetry.ExecutionTimeTaken().String()
+	analysis.PeakSamples = a.PeakSamples()
+	analysis.TotalSamples = a.TotalSamples()
 	for _, c := range a.Children {
 		analysis.Children = append(analysis.Children, processAnalysis(&c))
 	}