+++ pkg/dedup/iter.go
@@ -286,6 +286,7 @@ func newDedupSeriesIterator(a, b adjustableSeriesIterator) *dedupSeriesIterator 
 		b:        b,
 		lastT:    math.MinInt64,
 		lastIter: a,
+		useA:     true,
 		aval:     a.Next(),
 		bval:     b.Next(),
 	}