+++ pkg/store/proxy_test.go
@@ -2194,13 +2194,13 @@ func TestDedupRespHeap_Deduplication(t *testing.T) {
 
 	for _, tcase := range []struct {
 		responses []*storepb.SeriesResponse
-		testFn    func(responses []*storepb.SeriesResponse, h *dedupResponseHeap)
+		testFn    func(responses []*storepb.SeriesResponse, h *responseDeduplicator)
 		tname     string
 	}{
 		{
 			tname:     "edge case with zero responses",
 			responses: []*storepb.SeriesResponse{},
-			testFn: func(responses []*storepb.SeriesResponse, h *dedupResponseHeap) {
+			testFn: func(responses []*storepb.SeriesResponse, h *responseDeduplicator) {
 				testutil.Equals(t, false, h.Next())
 
 				callAtExpectPanic := func() {
@@ -2232,7 +2232,7 @@ func TestDedupRespHeap_Deduplication(t *testing.T) {
 					},
 				},
 			},
-			testFn: func(responses []*storepb.SeriesResponse, h *dedupResponseHeap) {
+			testFn: func(responses []*storepb.SeriesResponse, h *responseDeduplicator) {
 				testutil.Equals(t, true, h.Next())
 				resp := h.At()
 				testutil.Equals(t, responses[0], resp)
@@ -2274,7 +2274,7 @@ func TestDedupRespHeap_Deduplication(t *testing.T) {
 					},
 				},
 			},
-			testFn: func(responses []*storepb.SeriesResponse, h *dedupResponseHeap) {
+			testFn: func(responses []*storepb.SeriesResponse, h *responseDeduplicator) {
 				testutil.Equals(t, true, h.Next())
 				resp := h.At()
 				testutil.Equals(t, responses[0], resp)
@@ -2283,8 +2283,9 @@ func TestDedupRespHeap_Deduplication(t *testing.T) {
 		},
 	} {
 		t.Run(tcase.tname, func(t *testing.T) {
-			h := NewDedupResponseHeap(NewProxyResponseHeap(
+			h := NewResponseDeduplicator(NewProxyResponseLoserTree(
 				&eagerRespSet{
+					closeSeries:       func() {},
 					wg:                &sync.WaitGroup{},
 					bufferedResponses: tcase.responses,
 				},