+++ pkg/tracing/otlp/otlp.go
@@ -8,6 +8,8 @@ import (
 	"strconv"
 	"strings"
 
+	"go.opentelemetry.io/otel/attribute"
+
 	"github.com/thanos-io/thanos/pkg/tracing/migration"
 
 	"github.com/go-kit/log"
@@ -24,11 +26,14 @@ import (
 )
 
 const (
-	TracingClientGRPC string = "grpc"
-	TracingClientHTTP string = "http"
-	AlwaysSample      string = "alwayssample"
-	NeverSample       string = "neversample"
-	RatioBasedSample  string = "traceidratiobased"
+	TracingClientGRPC                  string = "grpc"
+	TracingClientHTTP                  string = "http"
+	AlwaysSample                       string = "alwayssample"
+	NeverSample                        string = "neversample"
+	TraceIDRatioBasedSample            string = "traceidratiobased"
+	ParentBasedAlwaysSample            string = "parentbasedalwayssample"
+	ParentBasedNeverSample             string = "parentbasedneversample"
+	ParentBasedTraceIDRatioBasedSample string = "parentbasedtraceidratiobased"
 )
 
 // NewOTELTracer returns an OTLP exporter based tracer.
@@ -67,26 +72,27 @@ func NewTracerProvider(ctx context.Context, logger log.Logger, conf []byte) (*tr
 	if err != nil {
 		logger.Log(err)
 	}
-	tp := newTraceProvider(ctx, processor, logger, config.ServiceName, sampler)
+	tp := newTraceProvider(ctx, processor, logger, config.ServiceName, config.ResourceAttributes, sampler)
 
 	return tp, nil
 }
 
-func newTraceProvider(ctx context.Context, processor tracesdk.SpanProcessor, logger log.Logger, serviceName string, sampler tracesdk.Sampler) *tracesdk.TracerProvider {
-	var (
-		r   *resource.Resource
-		err error
-	)
+func newTraceProvider(
+	ctx context.Context,
+	processor tracesdk.SpanProcessor,
+	logger log.Logger,
+	serviceName string,
+	attrs map[string]string,
+	sampler tracesdk.Sampler,
+) *tracesdk.TracerProvider {
+	resourceAttrs := make([]attribute.KeyValue, 0, len(attrs)+1)
 	if serviceName != "" {
-		r, err = resource.New(
-			ctx,
-			resource.WithAttributes(semconv.ServiceNameKey.String(serviceName)),
-		)
-	} else {
-		r, err = resource.New(
-			ctx,
-		)
+		resourceAttrs = append(resourceAttrs, semconv.ServiceNameKey.String(serviceName))
 	}
+	for k, v := range attrs {
+		resourceAttrs = append(resourceAttrs, attribute.String(k, v))
+	}
+	r, err := resource.New(ctx, resource.WithAttributes(resourceAttrs...))
 	if err != nil {
 		level.Warn(logger).Log("msg", "jaeger: detecting resources for tracing provider failed", "err", err)
 	}
@@ -106,10 +112,20 @@ func newTraceProvider(ctx context.Context, processor tracesdk.SpanProcessor, log
 func getSampler(config Config) (tracesdk.Sampler, error) {
 	switch strings.ToLower(config.SamplerType) {
 	case AlwaysSample:
-		return tracesdk.ParentBased(tracesdk.AlwaysSample()), nil
+		return tracesdk.AlwaysSample(), nil
 	case NeverSample:
+		return tracesdk.NeverSample(), nil
+	case TraceIDRatioBasedSample:
+		arg, err := strconv.ParseFloat(config.SamplerParam, 64)
+		if err != nil {
+			return tracesdk.TraceIDRatioBased(1.0), err
+		}
+		return tracesdk.TraceIDRatioBased(arg), nil
+	case ParentBasedAlwaysSample:
+		return tracesdk.ParentBased(tracesdk.AlwaysSample()), nil
+	case ParentBasedNeverSample:
 		return tracesdk.ParentBased(tracesdk.NeverSample()), nil
-	case RatioBasedSample:
+	case ParentBasedTraceIDRatioBasedSample:
 		arg, err := strconv.ParseFloat(config.SamplerParam, 64)
 		if err != nil {
 			return tracesdk.ParentBased(tracesdk.TraceIDRatioBased(1.0)), err